Improves participant management related to tasks
